fix: address all review feedback — schema validation, shared init, fail-closed
- Schema: zod refine rejects enabled=true without provider.use
- Init: extract initGuardrailsFromConfig to src/guardrails/init.ts (shared,
  not gateway-specific) so any startup path can call it
- Fail-closed: deny-all fallback installed immediately; replaced on load success
- Bootstrap: server-plugin-bootstrap.ts is now +2 lines (import + call)
- Integration: pi-tools.before-tool-call.ts has no async init on hot path
- Reasons: optional in type, defensive access in evaluateGuardrail
- Loader: local paths resolved relative to cwd
